refactor: LottoMachine Default Parameter 제거
