+++ README.md
@@ -80,3 +80,6 @@
     - 논리적인 오류가 아니면 예외를 던지지 말고 null을 반환한다.
     - 실패하는 경우가 복잡해서 null로 처리할 수 없으면 sealed class를 반환한다.
     
+### 피드백 반영
+
+- [x] LottoMachine의 디폴트 파라미터를 제거
