+++ src/main/kotlin/lotto/domain/model/LottoMachine.kt
@@ -1,10 +1,9 @@
 package lotto.domain.model
 
-import lotto.domain.service.LottoNumbersGenerator
-import lotto.domain.service.RandomLottoNumbersGenerator
+import lotto.domain.generator.LottoNumbersGenerator
 
 class LottoMachine(
-    private val generator: LottoNumbersGenerator = RandomLottoNumbersGenerator(),
+    private val generator: LottoNumbersGenerator,
 ) {
     fun generateLottoBundle(count: Int): LottoBundle? {
         if (count == 0) return null
